fix: PPMD/LZP/MM codec fixes for arm64 (64-bit) port
PPMD:
- PPMdType.h: Change DWORD from unsigned long (8 bytes on arm64) to
  uint32_t (always 4 bytes). The range coder in Coder.hpp requires
  32-bit DWORD for correct normalization behavior.
- Model.cpp: Replace DWORD pointer copy in CreateSuccessors with
  memcpy to correctly copy full 64-bit Stats pointer (was only
  copying low 32 bits, causing crash/garbage).
- SubAlloc.hpp: Added _pad field to MEM_BLK on 64-bit so
  sizeof(MEM_BLK) == sizeof(PPM_CONTEXT) == 20 bytes.

LZP:
- Common.h: Guard lb() against __builtin_clz(0) undefined behavior
  by returning 0 when n==0 (was returning 63 on arm64, causing
  impossibly large hash table allocation).
- GRZip/LZP.c: Fix hash table allocation from sizeof(uint32) to
  sizeof(uint8*) for correct pointer storage on 64-bit.

MM:
- mmdet.cpp: Fix _32bit_run/_32bit_diff_run from long* to int32_t*
  (long is 8 bytes on arm64, not 4).

+++ Compression/PPMD/PPMdType.h
@@ -13,12 +13,13 @@
 #ifdef FREEARC_WIN
 #include <windows.h>
 #else
+#include <stdint.h>
 typedef int   BOOL;
 #define FALSE 0
 #define TRUE  1
 typedef unsigned char  BYTE;
 typedef unsigned short WORD;
-typedef unsigned long  DWORD;
+typedef uint32_t       DWORD;
 typedef unsigned int   UINT;
 #endif
